requirements.txt
Flask==2.3.3
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
prometheus-client==0.17.1
sentry-sdk[flask]==1.38.0 